lane refactored - tested in sim. Works. Not tested on HW
param for sim & hw - laneDetn
