+++ autovalet_lane_detection/config/hardware_params.yaml
@@ -1,0 +1,12 @@
+hlsbounds:
+    lowerbound:
+        h: 12.0
+        l: 0.18
+        s: 0.90
+    upperbound:
+        h: 30.0
+        l: 1.00
+        s: 1.00
+lineparams:
+    minLineLength: 130
+    maxLineGap: 20
