+++ autovalet_lane_detection/config/sim_params.yaml
@@ -1,0 +1,12 @@
+hlsbounds:
+    lowerbound:
+        h: 25.0
+        l: 0.00
+        s: 0.30
+    upperbound:
+        h: 67.0
+        l: 0.90
+        s: 1.00
+lineparams:
+    minLineLength: 100
+    maxLineGap: 40
