+++ autovalet_lane_detection/launch/lane_detector.launch
@@ -1,5 +1,30 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no" ?>
 <launch>
-  	<node name="lane_detection" pkg="autovalet_lane_detection" type="autovalet_lane_detector.py" output="screen"/>
+    <arg name="simulation" default="true"/>
+
+    <!-- RUN LANEDETN WITH SIM PARAMETERS -->
+    <group if="$(arg simulation)">
+        <!-- Register depth image to rgb image -->
+        <node pkg="nodelet" type="nodelet" name="nodelet_manager" args="manager" />
+
+        <node pkg="nodelet" type="nodelet" name="nodelet1" args="load depth_image_proc/register nodelet_manager">
+            <remap from="depth/camera_info" to="/frontCamera/depth/camera_info"/>
+            <remap from="rgb/camera_info" to="/frontCamera/color/camera_info"/>
+            <remap from="depth/image_rect" to="/frontCamera/depth/image_raw"/>
+        </node>
+
+        <!-- Run Lane Detection Node-->
+        <node name="lane_detection" pkg="autovalet_lane_detection" type="autovalet_lane_detector.py" output="screen">
+            <rosparam file="$(find autovalet_lane_detection)/config/sim_params.yaml" command="load" />
+        </node>
+    </group>
+
+    <!-- RUN LANEDETN WITH HARDWARE PARAMETERS -->
+    <group unless="$(arg simulation)">
+        <!-- Run Lane Detection Node-->
+        <node name="lane_detection" pkg="autovalet_lane_detection" type="autovalet_lane_detector.py" output="screen">
+            <rosparam file="$(find autovalet_lane_detection)/config/hardware_params.yaml" command="load" />
+        </node>
+    </group>
 
 </launch>
