+++ autovalet_lane_detection/utils_legacy/mask_extractor.py
