+++ autovalet_navigation/launch/autovalet_nav.launch
@@ -1,3 +1,4 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no" ?>
 <launch>
 
    <arg name="simulation" default="true"/>
